// src/types/inmueble.types.ts
export interface Inmueble {
    c_inmuebles: number
    id_cliente: number
    tour: any
    activo_01: number
    referencia: string
    tipo: string
    pais: string
    provincia: string
    poblacion: string
    zona: string
    cp: string
    posicion_exacta_01: number
    latitud: any
    longitud: any
    direccion_visible_01: number
    direccion_completa_01: number
    direccion_tipo_via: string
    direccion_calle: string
    direccion_numero: string
    direccion_piso: string
    direccion_letra: string
    direccion_escalera: string
    precio_compra: any
    precio_alquiler: number
    precio_traspaso: any
    precio_alquiler_opcion_compra: any
    precio_ibi: any
    informacion_gestion: any
    visible: number
    opciones_inactivo: any
    n_habitaciones: number
    n_alcobas: any
    n_banos: number
    n_aseos: any
    m2_utiles: number
    m2_construidos: number
    m2_terreno: any
    estado_vivienda: number
    accesible_01: number
    categoria: any
    amueblado_01: number
    garaje_01: number
    calefaccion_01: number
    aire_acondicionado_01: number
    piscina_01: number
    jardin_01: number
    trastero_01: number
    ascensor_01: number
    terraza_01: number
    piso_banco_01: number
    vpo_01: number
    reservado_01: number
    exclusividad_01: number
    exclusividad_documento_url: string
    fecha_captacion: string
    fecha_fin_mandato: string
    aviso_fin_mandato: number
    referencia_catastral: string
    idufir: any
    rc_tomo: any
    rc_libro: any
    rc_folio: any
    rc_finca: any
    rc_registro: any
    eficiencia_energetica_tipo: string
    eficiencia_energetica_entramite_01: number
    eficiencia_energetica_fecvalid: any
    eficiencia_energetica_emisiones: number
    eficiencia_energetica_energia: number
    eficiencia_energetica_energia_tipo: string
    eficiencia_energetica_documento_url: any
    eficiencia_energetica_documento_visible_01: number
    eficiencia_energetica_n_expediente: any
    descripcion: any
    observaciones: any
    descripcion_ficha: string
    c_promociones: any
    nombre_comercializador: string
    email_comercializador: string
    telefono_comercializador: string
    gestion: number
    texto_reclamo: string
    estado: number
    zona_deportiva_01: number
    orientacion_solar: number
    gastos_comunidad: any
    altura_edificio: number
    etiqueta: any
    suelos: number
    carpinteria_exterior: number
    carpinteria_interior: number
    armarios_empotrados: number
    url: any
    a_consultar: number
    ano_construccion: number
    derramas_01: number
    hipoteca_01: number
    cantidad_pendiente_hipoteca: any
    banco_acreedor: any
    fecha_entrega_llaves: any
    llaves_disponibles_01: number
    cartel_requerido_01: number
    coste_plusvalia: any
    horario_visitas: any
    otros_datos: any
    honorarios_a_compartir_sin_iva: any
    honorarios_sin_iva: any
    created_at: string
    updated_at: string
    deleted_at: any
    adaptada_01: number
    fotos_inmueble?: FotosInmueble[]
    poblacion_inmueble?: PoblacionInmueble
    zona_inmueble?: ZonaInmueble
  }

  export interface FotosInmueble {
    c_fotos: number
    c_inmuebles: number
    id_category: any
    activo_01: any
    url: string
    original: string
    thumbnail: string
    principal: number
    descripcion_corta: any
    tipo: number
    orden: number
    marca: number
    redimensionada: any
    created_at: string
    updated_at: string
  }

  export interface PoblacionInmueble {
    id: number
    nombre: string
    codigo_pr: number
    activa: number
    principal: number
    created_at: string
    updated_at: string
  }

  export interface ZonaInmueble {
    id: number
    nombre: string
    codigo_po: number
    activa: number
    nombre_admin: any
    created_at: string
    updated_at: string
  }